modules/retrieval/text_classification/requirements.txt
absl-py==0.12.0
albumentations==0.5.2
alembic==1.5.8
anyio==2.2.0
appdirs==1.4.3
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
autonlp==0.2.6
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
Bottleneck==1.3.2
CacheControl==0.12.6
cachetools==4.2.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==8.0.1
cliff==3.7.0
cmaes==0.8.2
cmd2==1.5.0
colorama==0.4.3
colorlog==5.0.1
contextlib2==0.6.0
contextvars==2.4
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.0
distro==1.4.0
efficientnet-pytorch==0.7.1
entrypoints==0.3
filelock==3.0.12
google-auth==1.29.0
google-auth-oauthlib==0.4.4
greenlet==1.0.0
grpcio==1.37.0
html5lib==1.0.1
htmlmin==0.1.12
huggingface-hub==0.0.8
idna==2.8
ImageHash==4.2.0
imageio==2.9.0
imgaug==0.4.0
immutables==0.15
importlib-metadata==4.0.1
ipaddr==2.2.0
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.9.2
jupyter-server==1.6.4
jupyterlab==3.0.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.5.0
jupyterlab-widgets==1.0.0
kaggle==1.5.12
kiwisolver==1.3.1
lockfile==0.12.2
loguru==0.5.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
matplotlib-inline==0.1.2
missingno==0.4.2
mistune==0.8.4
msgpack==0.6.2
multimethod==1.4
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.3.0
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
optuna==2.7.0
packaging==20.3
pandas==1.1.5
pandas-profiling==3.0.0
pandocfilters==1.4.3
parso==0.8.2
pbr==5.6.0
pep517==0.8.2
pexpect==4.8.0
phik==0.11.2
pickleshare==0.7.5
Pillow==8.2.0
pkg-resources==0.0.0
prettytable==2.0.0
progress==1.5
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.8.1
pyparsing==2.4.6
pyperclip==1.8.2
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==4.0.1
pytoml==0.1.21
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
sacremoses==0.0.45
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
sentencepiece==0.1.95
Shapely==1.7.1
six==1.14.0
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.4.11
stevedore==3.3.0
tangled-up-in-unicode==0.1.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
terminado==0.9.4
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2020.9.3
timm==0.3.2
tokenizers==0.10.2
toml==0.10.2
tomlkit==0.7.0
torch==1.8.1
torch-tb-profiler==0.1.0
torchfile==0.1.0
torchnet==0.0.4
torchsummary==1.5.1
torchvision==0.9.1
tornado==6.1
tqdm==4.56.0
traitlets==4.3.3
transformers==4.6.1
typing-extensions==3.7.4.3
urllib3==1.25.8
visdom==0.1.8.9
visions==0.7.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.1
